Fix circular dependencies
